Extract portable/managed code to Array.cs for sharing with Mono and CoreCLR
Includes few missing GetLowerBound calls from CoreCLR version
